+++ actix-web-lab/src/extract.rs
@@ -5,6 +5,7 @@ pub type SharedData<T> = actix_web::web::Data<T>;
 
 #[allow(deprecated)]
 pub use crate::body_hash::BodyHash;
+#[allow(deprecated)]
 pub use crate::body_hmac::{BodyHmac, HmacConfig};
 pub use crate::json::{Json, DEFAULT_JSON_LIMIT};
 pub use crate::lazy_data::LazyData;